Add product creation functionality and refactor models
- Implemented `CreateProduct` method in `ProductController` for handling product creation via HTTP POST.
- Updated `Price` property types in view models to `decimal` for better monetary representation.
- Added `CreateProduct` method in `DataService` to interact with the API.
- Modified `CreateProductViewModel` to change property bindings and rename `Description` to `Details`.
- Refactored command handling with new `CreateProductCommand` and `CreateProductCommandHandler` returning `Guid`.
- Changed `TotalPrice` in `CartModel` to `decimal`.
- Updated `ShoppingCartItemModel` to use `decimal` for `Price` and `ProductVatRate`, adjusting calculations accordingly.
- Altered `IProductRepository` to return `Guid` from `Create` method, reflecting new product creation behavior.

+++ src/main/Hat.Mobile/DataViewModels/ShoppingCartItemViewModel.cs
@@ -26,11 +26,11 @@ namespace Hat.DataViewModels
         public string ProductDetails { get; set; }
         public string Name { get; set; }
         public string BrandName { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public double Qty { get; set; } = 1;
         public string Size { get; set; }
 
-        public double Amount { get; init; }
-        public double Vat { get; init; }
+        public decimal Amount { get; init; }
+        public decimal Vat { get; init; }
     }
 }
