+++ src/main/libs/Hat.Domain/Models/ShoppingCartItemModel.cs
@@ -8,7 +8,7 @@
         public Guid ProductId { get; set; }
 
 
-        public double ProductVatRate { get; set; } // Optional, if you want to track VAT separately
+        public decimal ProductVatRate { get; set; } // Optional, if you want to track VAT separately
         public string ProductName { get; set; }
 
         public string ProductColor { get; set; } // Optional, if you want to track color separately
@@ -18,14 +18,14 @@
 
         public string ProductDetails { get; set; }
         public int Qty { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public double Vat { get; set; } = 0.2; // 20% VAT by default
         public string Size { get; set; }
         #endregion
 
         #region rich
-        public double Amount => Price * Qty;
-        public double VatAmount => Amount * ProductVatRate;
+        public decimal Amount => Price * Qty;
+        public decimal VatAmount => Amount * ProductVatRate;
 
         #endregion
     }
